Updated Readme again. [It's really coming along]
Added a table of contents, Updated the contributions, and also added the naming conventions of classes fields and the commenting methods we are using for Methods. There are also headers now for two other topics to be implemented soon. One being a bug report information and the other is the building of the Mod.